fase-4/tech-challenge/requirements.txt
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
decorator==5.2.1
google-api-core==2.29.0
google-auth==2.48.0
google-cloud-speech==2.36.1
googleapis-common-protos==1.72.0
grpcio==1.78.0
grpcio-status==1.78.0
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
moviepy==2.2.1
numpy==2.0.2
pillow==11.3.0
proglog==0.1.12
proto-plus==1.27.1
protobuf==6.33.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
python-dotenv==1.2.1
requests==2.32.5
rsa==4.9.1
tqdm==4.67.3
typing_extensions==4.15.0
urllib3==2.6.3
